backend-users/requirements.txt
# requirements.txt
fastapi 
sqlmodel 
uvicorn[standard]
python-dotenv
alembic
databases[postgresql]
websocket
psycopg2-binary
asyncpg 
python-dotenv
kafka-python
pydantic-settings

numpy
pandas
# opencv-python>=4.1.1
Pillow>=7.1.2
# python pep8 
autopep8